refactor: move helper utilities
